Chapter 20 - Section 3: Graceful shutdown
Use the Drop trait to implement a custom behaviour to shutdown all components
when main() blocks finishes its execution. Use Option<..> and drop(..) to acomplish this
objective. For multiproducer-single consumer pattern, as soon as the sender is dropped all the receivers will get an error when trying to process the message, use this as a signal to shutdown the internal workers